Using request module
Since we have to support redirects from HTTP to HTTPS, it is easier
to use the request module and to send a completed request instead
of streaming the data.

This makes the recording proxy simpler but possibly less performant
as the request must finish before the targetRequest is started.

WIP: Added failing test for HTTPS redirect

WIP: Added _createReverseProxy

Passing target into _createReverseProxy