Resend HTTP request if ECONNRESET
If the server is throttling requests, it will listen to the port,
but close the socket with an ECONNRESET (not reading data).

The recording proxy can safely resend the request after a delay.

This helps "bursty" traffic get spread out but will not
help if the incoming requests are sustained above the server rate.